更改imread返回的数据类型为内置的int型. 多加了一层保险应该不会出现溢出问题了

+++ __init__.py
@@ -21,7 +21,7 @@ def abs_fft(array):
 
 
 def imread(img_path):
-    return cv.imread(img_path, cv.IMREAD_GRAYSCALE | cv.IMREAD_ANYDEPTH)
+    return cv.imread(img_path, cv.IMREAD_GRAYSCALE | cv.IMREAD_ANYDEPTH).astype(int)
 
 
 def read_csv(path):
